requirements.txt
pytest==7.4.3
requests==2.31.0
requests-toolbelt==1.0.0
allure-pytest==2.13.2
Faker==37.8.0
